rename component for clarity
